% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allperm.R
\name{allperm}
\alias{allperm}
\title{Calculate all values for two-stage Mann-Whitney-Wilcoxon statistic, with frequencies}
\usage{
allperm(m1, n1, m2, n2, delta)
}
\arguments{
\item{m1}{Number in group 1, first evaluation}

\item{n1}{Number in group 2, first evaluation}

\item{m2}{Number in group 1, second evaluation}

\item{n2}{Number in group 2, second evaluation}

\item{delta}{Under normal unit variance model for responses, separation between means in groups.}
}
\value{
Matrix with three columns, reprsenting values of statistic at time 1, values at time 2, and count.
}
\description{
Calculate all values for two-stage Mann-Whitney-Wilcoxon statistic, with frequencies
}
\examples{
allperm(2,2,2,2,0)

}
