% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checktest.R
\name{checktest}
\alias{checktest}
\title{Calculates values associated with testing using two-stage Mann-Whitney-Wilcoxon statistic.}
\usage{
checktest(
  m1,
  n1,
  m2,
  n2,
  delta,
  quantiles = c(0.98, 0.95),
  docheck = TRUE,
  ctl = FALSE
)
}
\arguments{
\item{m1}{Number in group 1, first evaluation}

\item{n1}{Number in group 2, first evaluation}

\item{m2}{Number in group 1, second evaluation}

\item{n2}{Number in group 2, second evaluation}

\item{delta}{Under normal unit variance model for responses, separation between means in groups.}

\item{quantiles}{Quantiles for first stage and overall 1-size.}

\item{docheck}{Flag controling whether checking vs a table of the joint distribution is done.  Default is true.}

\item{ctl}{Flag controling whether simulation is used to get emprical distribution.}
}
\value{
List with componets delta, quantiles at which critical values ought to be calculated, m1, n2, m2, n2 from input,normal-approximation critical values, Cornish-Fisher critical values,true size of normal theory test, true size of Cornish-Fisher test, true power under delta, Edgworth approximation to power under normal critical values, Edgeworth approximation to power under Cornish-Fisher critical values.
}
\description{
Calculates values associated with testing using two-stage Mann-Whitney-Wilcoxon statistic.
}
\examples{
checktest(2,2,2,2,1)

}
