% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_est}
\alias{compare_est}
\title{Compare Estimates of Common Parameters}
\usage{
compare_est(...)
}
\arguments{
\item{...}{objects (lavaan or twostage)}
}
\value{
A data frame with estimates and SEs from both models for common model parameters
in the style of \code{parameterestimates()} of \code{lavaan}
}
\description{
Compare Estimates of Common Parameters
}
