% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIM.R
\name{compmodel_base}
\alias{compmodel_base}
\title{Title}
\usage{
compmodel_base(compmodel, exog_cov = TRUE)
}
\arguments{
\item{compmodel}{Composite model syntax for which to create a baseline model}

\item{exog_cov}{Should exogenous variables in the compmodel be correlated in the baseline model? If FALSE, all variables in the baseline model will be orthogonal.}
}
\value{
Baseline model lavaan syntax
}
\description{
Title
}
\examples{
compmodel<-"C1 ~ C2 + C3"
basemodel <- compmodel_base(compmodel) #default is to covary C2 and C3
basemodel1 <- compmodel_base(compmodel,exog_cov=FALSE) #all three composites are orthogonal

}
