% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIM.R
\name{compmodel_sat}
\alias{compmodel_sat}
\title{H1 (Saturated) Model Syntax for the Composites Model}
\usage{
compmodel_sat(compmodel)
}
\arguments{
\item{compmodel}{Composite model syntax for which to create a baseline model}
}
\value{
H1 (Saturated) model lavaan syntax
}
\description{
H1 (Saturated) Model Syntax for the Composites Model
}
\examples{
compmodel<-"C1 ~ C2 + C3"
satmodel <- compmodel_sat(compmodel)

}
