% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doperm.R
\name{doperm}
\alias{doperm}
\title{Calculate an MC values for two-stage Mann-Whitney-Wilcoxon statistic, with frequencies}
\usage{
doperm(m1, m2, n1, n2, nsamp = 1000)
}
\arguments{
\item{m1}{Number in group 1, first evaluation}

\item{m2}{Number in group 1, second evaluation}

\item{n1}{Number in group 2, first evaluation}

\item{n2}{Number in group 2, second evaluation}

\item{nsamp}{Number of MC samples.}
}
\value{
Table of relative frequencies.
}
\description{
Calculate an MC values for two-stage Mann-Whitney-Wilcoxon statistic, with frequencies
}
\examples{
doperm(2,2,2,2)

}
