% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmeasures_PIM.R
\name{fitMeasures_pim}
\alias{fitMeasures_pim}
\title{PIM fit measures based on correct baseline and saturated models}
\usage{
fitMeasures_pim(C, compmodel, fit_pim, data = data, exog_cov = TRUE)
}
\arguments{
\item{C}{A matrix of 0s and 1s, where rows are composites and columns are
components}

\item{compmodel}{The lavaan model for the composites}

\item{fit_pim}{A lavaan object for the fitted PIM model}

\item{data}{Data file for the components (items)}

\item{exog_cov}{Should exogenous variables in the baseline model for composites be correlated?}
}
\value{
a fitMeasures vector of lavaan, truncated to include only indices that can be interpreted (are already correct or have been adjusted)
}
\description{
Returns PIM fit measures, where incremental fit indices have been adjusted to be based on a custom PIM baseline model, and where SRMR indices have been adjusted to be based on a custom H1 (saturated) model
Indices that have not been adjusted and should not be interpreted for PIM are omitted (for now)
RMSEA is already correct
}
\examples{
# An example using the first 18 variables in the simulated
# dataset misdata_mcar20 with 20\% missing data on about half the variables
# C1 - C6 are parcels formed using three variables each (in order)

library(lavaan)
misdata1 <- misdata_mcar20[, 1:18]

# composite sub-model
mod1 <- "
 F1 =~ C1 + C2 + C3
 F2 =~ C4 + C5 + C6
 F2 ~ F1
 F2 ~~ F2
 F1 ~~ F1"
#manual computation for C to avoid user input
cnames<-lavNames(mod1) #this is a lavaan function
C <- matrix(0,nrow=length(cnames),ncol=length(colnames(misdata1)))
colnames(C)<-colnames(misdata1)
rownames(C)<-cnames
C[1,1:3]<-1
C[2,4:6]<-1
C[3,7:9]<-1
C[4,10:12]<-1
C[5,13:15]<-1
C[6,16:18]<-1

pim_mod1 <- PIM_syntax(mod1,C)
fit_pim <- lavaan::lavaan(pim_mod1, data=misdata1)

fitMeasures_pim(C = C, compmodel = mod1, fit_pim = fit_pim,
data = misdata1, exog_cov = TRUE)

}
