% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{GaspardEtal2015mathUtility}
\alias{GaspardEtal2015mathUtility}
\title{GaspardEtal2015mathUtility}
\format{
The order of variables in the data set is: 1.	id, 2.	school, 3.	class, 4.	sex, 5.	math
6.	ca, 7.	isei_m, 8.	isei_d, 9.	edu_m, 10.	edu_d, 11.	voc_m, 12.	voc_d, 13.	book
14.	income, 15.	y1 - y12

The variables isei_m, isei_d, edu_m, edu_d, voc_m, voc_d, book, and income are z-standardized.
More Information about the data and the variables can be found in Rose et al. (2019).
}
\source{
\url{https://osf.io/f3kzb/}
}
\usage{
GaspardEtal2015mathUtility
}
\description{
A dataset provided by Rose et al (2019). The dataset contains 26 variables.
It has been modified by adding variable names and changing missing data codes
from -99 to NA.
}
\references{
Rose, N., Wagner, W., Mayer, A., & Nagengast, B. (2019). Model-based manifest and latent
composite scores in structural equation models. Collabra: Psychology, 5(1), Article 9.
https://doi.org/10.1525/collabra.143
}
\keyword{datasets}
