% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.R
\name{inside}
\alias{inside}
\title{Calculate all values for two-stage Mann-Whitney-Wilcoxon statistic, with frequencies}
\usage{
inside(m1, m2, n1, n2, delta, alpha, xvn)
}
\arguments{
\item{m1}{Number in group 1, first evaluation}

\item{m2}{Number in group 1, second evaluation}

\item{n1}{Number in group 2, first evaluation}

\item{n2}{Number in group 2, second evaluation}

\item{delta}{Under normal unit variance model for responses, separation between means in groups.}

\item{alpha}{Two component real vector giving target probabilities at each stage.}

\item{xvn}{Two component real vector giving normal deviates associated with alpha.  Note that this required knowledge of the correlation, outside of this function.}
}
\value{
List with components MC size of normal critical value, MC size of CF critical value, Edgeworth 1-size of normal critical value, Edgeworth 1-size of CF critical value, Normal type 2 error for corrected critical values, Edgeworth type 2 error for corrected critical values, Monte Carlo type 2 error for uncorrected critical values, Monte Carlo type 2 error for corrected critical values.
}
\description{
Calculate all values for two-stage Mann-Whitney-Wilcoxon statistic, with frequencies
}
\examples{
inside(m1=2,n1=2,m2=2,n2=2,delta=1,alpha=c(.98,.95),xvn=c(-2.3263482318637529,-2.0863447264627704))

}
