% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{parameterEstimates_ts}
\alias{parameterEstimates_ts}
\alias{parameterestimates_ts}
\title{Display Twostage Parameter Estimates and SEs}
\usage{
parameterEstimates_ts(object, naive.se = TRUE)

parameterestimates_ts(object, naive.se = TRUE)
}
\arguments{
\item{object}{Object of class twostage}

\item{naive.se}{Whether naive standard errors from Stage 2 should be included}
}
\value{
Dataframe with TS estimates and SEs
}
\description{
Display Twostage Parameter Estimates and SEs
}
