% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIM.R
\name{PIM.multi}
\alias{PIM.multi}
\title{PIM.multi() function creates the second part of the PIM model syntax}
\usage{
PIM.multi(C)
}
\arguments{
\item{C}{A matrix of 0s and 1s, where rows are composites and columns are components}
}
\value{
A string with the second part of the PIM model syntax
}
\description{
PIM.multi() function creates the second part of the PIM model syntax
}
\examples{
#C1 is the sum of Y1, Y2, and Y3
#C2 is the sum of Y4, Y5, and Y6
#C3 is Y7
C<-matrix(0,nrow=3,ncol=7)
C[1,1:3]<-1
C[2,4:6]<-1
C[3,7]<-1
rownames(C)<-c("C1","C2","C3")
colnames(C)<-c("Y1","Y2","Y3","Y4","Y5","Y6","Y7")
PIM.multi(C)
}
