% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIM.R
\name{PIM_syntax}
\alias{PIM_syntax}
\title{Creates the lavaan syntax for a PIM (Pseudo-Indicator Model)
the names of the composite variables should match row names of the C matrix}
\usage{
PIM_syntax(C, compmodel)
}
\arguments{
\item{C}{A matrix of 0s and 1s, where rows are composites and columns are components}

\item{compmodel}{A string with the lavaan model for composites}
}
\value{
A string with the full PIM model syntax
}
\description{
Creates the lavaan syntax for a PIM (Pseudo-Indicator Model)
the names of the composite variables should match row names of the C matrix
}
\examples{
#C1 is the sum of Y1, Y2, and Y3
#C2 is the sum of Y4, Y5, and Y6
#C3 is Y7
C<-matrix(0,nrow=3,ncol=7)
C[1,1:3]<-1
C[2,4:6]<-1
C[3,7]<-1
rownames(C)<-c("C1","C2","C3")
colnames(C)<-c("Y1","Y2","Y3","Y4","Y5","Y6","Y7")
compmodel<-"C1 ~ C2 + C3"
PIM_syntax(C,compmodel)

}
\references{
Rose, N., Wagner, W., Mayer, A., & Nagengast, B. (2019). Model-based manifest and latent composite scores in structural equation models. Collabra: Psychology, 5(1), Article 9. https://doi.org/10.1525/collabra.143
}
