% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIM.R
\name{PIM_syntax_base}
\alias{PIM_syntax_base}
\title{Baseline Model Syntax for the PIM}
\usage{
PIM_syntax_base(C, compmodel_base)
}
\arguments{
\item{C}{A matrix of 0s and 1s, where rows are composites and columns are
components}

\item{compmodel_base}{A string with lavaan baseline (null) model for composites}
}
\value{
A string with the full PIM baseline model lavaan syntax
}
\description{
Baseline Model Syntax for the PIM
}
\examples{
# C1 is the sum of Y1, Y2, and Y3
# C2 is the sum of Y4, Y5, and Y6
# C3 is Y7
C <- matrix(0, nrow = 3, ncol = 7)
C[1, 1:3] <- 1
C[2, 4:6] <- 1
C[3, 7] <- 1
rownames(C) <- c("C1", "C2", "C3")
colnames(C) <- c("Y1", "Y2", "Y3", "Y4", "Y5", "Y6", "Y7")
compmodel <- "C1 ~ C2 + C3"
basemodel <- compmodel_base(compmodel)
PIM_model_base <- PIM_syntax_base(C = C, compmodel = basemodel)

}
