% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.SummaryTwostage}
\alias{print.SummaryTwostage}
\title{Print method for SummaryTwostage objects (S3)}
\usage{
\method{print}{SummaryTwostage}(x, ...)
}
\arguments{
\item{x}{An object of class 'SummaryTwostage'}

\item{...}{Additional arguments (currently not used)}
}
\description{
Print method for SummaryTwostage objects (S3)
}
