% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstdcum.R
\name{rstdcum}
\alias{rstdcum}
\title{Calculate standardized joint cumulants for sequential Mann Whitney wilcoxon statistic}
\usage{
rstdcum(m1, m2, n1, n2, delta)
}
\arguments{
\item{m1}{Number in group 1, first evaluation}

\item{m2}{Number in group 1, second evaluation}

\item{n1}{Number in group 2, first evaluation}

\item{n2}{Number in group 2, second evaluation}

\item{delta}{Under normal unit variance model for responses, separation between means in groups.}
}
\value{
List with components kk1 through kk4, with order 1 through 4 order joint cumulants, with 1 through 4 dimensions respectively.
}
\description{
Calculate standardized joint cumulants for sequential Mann Whitney wilcoxon statistic
}
\examples{
rstdcum(2,2,2,2,0)

}
