% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{samplesize}
\alias{samplesize}
\title{Calculates approximate sample size for two-stage Mann-Whitney-Wilcoxon statistic.}
\usage{
samplesize(
  omega,
  lambda11 = 0.25,
  lambda12 = 0.25,
  lambda21 = 0.25,
  lambda22 = 0.25,
  alpha2 = 0.05,
  beta = 0.2,
  round = TRUE
)
}
\arguments{
\item{omega}{Alternative P(Y>X)-.5}

\item{lambda11}{Proportion in group 1, first evaluation}

\item{lambda12}{Proportion in group 1, second evaluation}

\item{lambda21}{Proportion in group 2, first evaluation}

\item{lambda22}{Proportion in group 2, second evaluation}

\item{alpha2}{Overall test size}

\item{beta}{Type II error rate.}
}
\value{
group sizes
}
\description{
Calculates approximate sample size for two-stage Mann-Whitney-Wilcoxon statistic.
}
