% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_summary_and_print_methods.R
\name{show,SummaryTwostage-method}
\alias{show,SummaryTwostage-method}
\title{Show method for 'SummaryTwostage' objects}
\usage{
\S4method{show}{SummaryTwostage}(object)
}
\arguments{
\item{object}{An object of class 'SummaryTwostage'}
}
\description{
Show method for 'SummaryTwostage' objects
}
