% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,twostage-method}
\alias{show,twostage-method}
\title{Display method for twostage objects}
\usage{
\S4method{show}{twostage}(object)
}
\arguments{
\item{object}{A twostage object}
}
\description{
Display method for twostage objects
}
