% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmeasures_PIM.R
\name{srmr_mplus_pim}
\alias{srmr_mplus_pim}
\title{Title}
\usage{
srmr_mplus_pim(lavobject, lavobject_sat, compmodel)
}
\arguments{
\item{lavobject}{Estimated PIM (H0)}

\item{lavobject_sat}{Estimated saturated PIM (H1)}

\item{compmodel}{Composites Model}
}
\value{
A list of SRMR values, without means and with means
}
\description{
Title
}
\examples{
# make a non-boring example, here H0 is saturated
library(lavaan)
misdata_1to7 <- misdata_mcar20[, 1:7] # first 7 vars, for illustration
# C1 is the sum of Y1, Y2, and Y3
# C2 is the sum of Y4, Y5, and Y6
# C3 is Y7
C <- matrix(0, nrow = 3, ncol = 7)
C[1, 1:3] <- 1
C[2, 4:6] <- 1
C[3, 7] <- 1
rownames(C) <- c("C1", "C2", "C3")
colnames(C) <- c("Y1", "Y2", "Y3", "Y4", "Y5", "Y6", "Y7")
compmodel <- "C1 ~ C2 + C3"
PIM_model <- PIM_syntax(compmodel = compmodel, C = C)
PIM_model_sat <- PIM_syntax_sat(compmodel = compmodel,C = C)
fit_pim <- lavaan(PIM_model, data = misdata_1to7)
fit_pim_sat <- lavaan(PIM_model_sat, data = misdata_1to7)
srmr <- srmr_mplus_pim(fit_pim, fit_pim_sat, compmodel)
}
