% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{stage0}
\alias{stage0}
\title{stage0}
\usage{
stage0(data, model)
}
\arguments{
\item{data}{A data file with components (items) to be assigned to composites.
It should not contain other variables (or else feed in only selected columns as data).}

\item{model}{A lavaan model for the composites. It is only used to extract composite names
via lavNames().}
}
\value{
A matrix C with rownames set to composite names, and columnnames set to
component (item) names. A non-zero value in each row corresponds to the assignment of that
component to that composite, with the non-zero value as the weight. Current options for weights
are 1 for unit-weighted and 1/pj for average-weighted, where pj is the number of components in that
composite.
}
\description{
This is the prep stage of TSML. This function uses p column names of data and k composite variable names from the lavaan
model to create a k x p matrix C, with nonzero elements representing the assignment
of components to composites. Which elements are nonzero is determined by presenting
a menu of composite names to the user for each component. What the weights are is
determined by presenting the user with the menu to create composites as sum scores
or average scores.
}
\examples{
\dontrun{
# TPB Model for Composites (Full mediation)
tpbmod<-'
INTALL ~ ATTALL + PBCALL + NORMALL
BEH ~ INTALL'

stage0(tpbdata, tpbmod)
}

# With appropriate selections, this should result in the message:
#Your composites are made up of the following components:
#INTALL :  INT1 INT2 INT3
# BEH :  BEH
#ATTALL :  AT1CPU AT2CPU AT3CPU AT4CPU AT5CPU AT6CPU AT7CPP AT8CPP AT9CPP AT10CPP AT11CPP
#PBCALL :  PBC1 PBC2 PBC3
#NORMALL :  NORS1 NORS2 NORS3
#If this is not correct, start over!

}
