% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{stage1}
\alias{stage1}
\title{stage1}
\usage{
stage1(data, runcommand = NULL)
}
\arguments{
\item{data}{A datafile with components (items).
It should not contain any other variables.}

\item{runcommand}{Additional arguments to pass to lavaan}
}
\value{
A list with 3 objects: mhb (estimated means),
shb (estimated covariance matrix), ohb (estimated asymptotic covariance matrix
of mhb and shb elements)
}
\description{
Stage1 of TSML: fits a saturated model to the items with missing data
}
\examples{
\dontrun{
out_s1<-stage1(misdata_mcar20)

#as tpbdata has no missing data, running stage1 with expected information
#will result in TSML matching regular ML
out_s1<-stage1(tpbdata, runcommand="information='expected'")
}

}
