% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{stage1a}
\alias{stage1a}
\title{stage1a}
\usage{
stage1a(S1.output, C)
}
\arguments{
\item{S1.output}{The output of stage1, a list with three objects (means, covariance matrix,
and asymptotic covariance matrix of components)}

\item{C}{The matrix of component-composite assignments that is the output of stage0
(or user-supplied)}
}
\value{
A list with 3 objects: mhd (estimated means),
shd (estimated covariance matrix), ohd (estimated asymptotic covariance matrix
of mhb and shb elements) for the composites
}
\description{
Stage1a of TSML: Converts the means, covariance matrix, and asymptotic covariance matrix
of components to the corresponding quantities for the composites
}
\examples{

#an example using the first 18 variables in the simulated dataset misdata_mcar20
#reduce model size

library(lavaan)
misdata1<-misdata_mcar20[,1:18]

# composite sub-model
mod1 <- '
 F1 =~ C1 + C2 + C3
 F2 =~ C4 + C5 + C6
 F2 ~ F1
 F2 ~~ F2
 F1 ~~ F1'
#manual computation for C (stage0) to avoid user input
cnames<-lavNames(mod1) #this is a lavaan function
C <- matrix(0,nrow=length(cnames),ncol=length(colnames(misdata1)))
colnames(C)<-colnames(misdata1)
rownames(C)<-cnames
C[1,1:3]<-1
C[2,4:6]<-1
C[3,7:9]<-1
C[4,10:12]<-1
C[5,13:15]<-1
C[6,16:18]<-1

out_s1<-stage1(misdata1)
out_s1a<-stage1a(out_s1,C)

}
