% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{stage2}
\alias{stage2}
\title{stage2}
\usage{
stage2(
  S1a.output,
  compmodel,
  runcommand2 = NULL,
  lavaan_function = c("sem", "lavaan", "cfa", "growth", "sam")
)
}
\arguments{
\item{S1a.output}{Output from stage1a, a list with four objects (shd, mhd, ohd, N)}

\item{compmodel}{The lavaan model for the composites}

\item{runcommand2}{Additional arguments to pass to lavaan}

\item{lavaan_function}{Which lavaan function to use to fit the model}
}
\value{
An object of class \code{lavaan}
}
\description{
Performs stage2 of TSML. Fits the model to the estimated vector of means and the estimated
covariance matrix for the composites using complete data routines, adjusts naive
standard errors to obtain robust TS standard errors (for normal data), computes
the residual-based test statistic (for normal data)
}
\examples{

# an example using the first 18 variables in the simulated dataset misdata_mcar20
# reduce model size
library(lavaan)
misdata1 <- misdata_mcar20[, 1:18]

# composite sub-model
mod1 <- "
 F1 =~ C1 + C2 + C3
 F2 =~ C4 + C5 + C6
 F2 ~ F1
 F2 ~~ F2
 F1 ~~ F1"
#manual computation for C (stage0) to avoid user input
cnames<-lavNames(mod1) #this is a lavaan function
C <- matrix(0,nrow=length(cnames),ncol=length(colnames(misdata1)))
colnames(C)<-colnames(misdata1)
rownames(C)<-cnames
C[1,1:3]<-1
C[2,4:6]<-1
C[3,7:9]<-1
C[4,10:12]<-1
C[5,13:15]<-1
C[6,16:18]<-1

out_s1 <- stage1(misdata1)
out_s1a <- stage1a(out_s1,C)
out_s2 <- stage2(out_s1a, compmodel = mod1, runcommand2="mimic='EQS'")
}
