% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{summary,twostage-method}
\alias{summary,twostage-method}
\title{Summary method for twostage objects}
\usage{
\S4method{summary}{twostage}(object, ...)
}
\arguments{
\item{object}{An object of class "twostage".}

\item{...}{Additional arguments (currently not used).}
}
\value{
The summary of the "twostage" object.
}
\description{
Provides a summary for objects of class "twostage".
}
