% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{summary.twostage}
\alias{summary.twostage}
\title{Summary method for twostage function}
\usage{
summary.twostage(object, ...)
}
\arguments{
\item{object}{An object of class 'twostage'.}

\item{...}{Additional arguments passed to the summary method (none so far)}
}
\description{
Summary method for twostage function
}
