% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabline.R
\name{tabline}
\alias{tabline}
\title{Convert output from function checktest assessing various aspects in a two-stage design into a line of the table in  the manuscript}
\usage{
tabline(cto, latex = TRUE)
}
\arguments{
\item{cto}{output from checktest}

\item{latex}{flag governing whether the line is formatted in a way that lets it be placed directly into the paper.}
}
\value{
The table line
}
\description{
Convert output from function checktest assessing various aspects in a two-stage design into a line of the table in  the manuscript
}
