% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tpbdata.R
\docType{data}
\name{tpbdata}
\alias{tpbdata}
\title{tpbdata}
\format{
A matrix with 108 rows and 21 columns:

\describe{
\item{PBC1, PBC2, PBC3}{Perceived Behavioral Control items; e.g.,
"How much control do you have over sticking to a diet for the next four weeks?"}
\item{NORS1, NORS2, NORS3}{Subjective Norms items; "List three most important
persons in your life and indicate how much each would (dis)approve"}
\item{AT1CPU-AT6CPU,AT7CPP-AT11CPP}{Semantic differential items measuring attitudes;
"I think that to keep to a diet in order to decrease my body weight during the next
four weeks is: Useless - Useful, Ineffective - Effective,
Disadvantageous - Advantageous, Stupid - Intelligent, Punishing - Rewarding, Foolish - Wise,
Unpleasant - Pleasant, Joyless - Joyful, Boring - Exciting, Unattractive - Attractive,
Unenjoyable - Enjoyable." (7 pt scale)}
\item{INT1, INT2, INT3}{Intention items; e.g., "I will expend effort to stick to a diet in the next four weeks".}
\item{BEH}{Self-report of behavior, collected over the phone four weeks later.}
}
}
\usage{
tpbdata
}
\description{
A real dataset collected to test the theory of planned behavior. Data were collected from a
sample of N = 108 college students who had the goal of reducing or maintaining their
current body weight. There are 11 semantic differential items measuring attitudes
toward dieting (6 tapping the cognitive and 5 tapping the affective component
of the attitudes), 3 items measures Norms, 3 items measuring perceived behavioral control,
3 items measuring intentions, and 1 item measuring self-report of behavior.
For more details, see Perugini and Bagozzi (2001), and Savalei and Bentler (2006).
}
\references{
Perugini, M., and Bagozzi, R. P. (2001). The role of desires and anticipated emotions
in goal-directed behaviours: Broadening and deepening the theory of planned behaviour.
British Journal of Social Psychology, 40(1), 79-98. https://doi.org/10.1348/014466601164704

Savalei, V., and Bentler, P. M. (2006). Structural equation modeling.
In Handbook of Statistics, Volume 26 (pp. 330-364). Elsevier.
}
\keyword{datasets}
