% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{twostage-class}
\alias{twostage-class}
\title{Two-stage maximum likelihood estimation}
\description{
Objects returned by the twostage function, inheriting from lavaan
}
\section{Slots}{

\describe{
\item{\code{twostage}}{List containing TSML-specific information}
}}

\section{Methods}{

\describe{
\item{show}{Modifies lavaan's default display with Stage 1 and Stage 2 convergence info and TS test statistic information}
\item{summary}{Summary of parameter estimates with naive and TSML standard errors, and test statistic summary output}
\item{fitMeasures}{Fit measures output of lavaan is extended by adding two new elements (for now): residual-based TSML chi-square and its p-value}
}
}

