\name{TwoStage-package}
\alias{TwoStage-package}
\alias{TwoStage}
\docType{package}
\title{
FDR control for multiple hypothesis tests of two-stage designs.
}
\description{
The TwoStage package implement hidden-Markov-model-based FDR control procedure as described in Wang et al (2014). It includes Zehetmayer's approach, and Sarkar's p value combination methods. 
}
\details{
\tabular{ll}{
Package: \tab TwoStage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-28\cr
License: \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Xiaoshan Wang <xwang at forsyth dot org>
}
\references{
X Wang, J. Starr, F. Jorge (2014) Two-stage design optimizes power and false discovery rates for microbiome research. 

SK Sarkar, J. Chen, W. Guo (2013). Multiple Testing in a Two-Stage Adaptive Design with Combination Tests Controlling FDR. J. Am. Stat. Assoc. 108: 1385-1401.

W. Sun and T. Cai (2009). Large-scale multiple testing under dependence. J. R.Stat. Soc B. 71, 393-424.
}


