% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSML.R
\name{twostage}
\alias{twostage}
\title{Title}
\usage{
twostage(data, model)
}
\arguments{
\item{data}{Your data file with components (items)}

\item{model}{Your model for the composites}
}
\value{
A list with several objects
}
\description{
Title
}
\examples{
#To be added later

}
\references{
Savalei, V., & Rhemtulla, M. (2017). Normal theory two-stage ML estimator when data are missing at the item level. Journal of Educational and Behavioral Statistics, 42(1), 1-27. https://doi.org/10.3102/1076998617694880
}
