% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilding.R
\name{wilding}
\alias{wilding}
\title{Calculate mass function for two-stage Mann-Whitney-Wilcoxon statistic}
\usage{
wilding(u1, u2, m1, n1, m2, n2)
}
\arguments{
\item{u1}{Statistic at first evaluation}

\item{u2}{Statistic at second evaluation}

\item{m1}{Number in group 1, first evaluation}

\item{n1}{Number in group 2, first evaluation}

\item{m2}{Number in group 1, second evaluation}

\item{n2}{Number in group 2, second evaluation}
}
\value{
Mass function
}
\description{
Calculate mass function for two-stage Mann-Whitney-Wilcoxon statistic
}
\examples{
wilding(1,1,2,2,2,2)

}
