% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_est}
\alias{compare_est}
\title{Compare TSML and PIM Estimates and Fit}
\usage{
compare_est(fitted_ts, fitted_pim)
}
\arguments{
\item{fitted_ts}{A twostage object}

\item{fitted_pim}{A lavaan object for a PIM model}
}
\value{
A data frame with estimates and SEs from both models for the
composite-level model parameters (i.e., excluding item-level parameters of PIM),
in the style of \code{parameterestimates()} of \code{lavaan}, merged into one, with key.
}
\description{
Compare TSML and PIM Estimates and Fit
}
