% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{parameterEstimates_ts}
\alias{parameterEstimates_ts}
\alias{parameterestimates_ts}
\title{Parameter Estimates for twostage objects}
\usage{
parameterEstimates_ts(object, naive.se = TRUE, ...)

parameterestimates_ts(object, naive.se = TRUE, ...)
}
\arguments{
\item{object}{An object of class twostage}

\item{naive.se}{(logical) Whether to print naive Stage 2 SEs}

\item{...}{Other arguments to pass to parameterEstimates of lavaan}
}
\value{
Object of class twostage_parameterEstimates with TS SEs
}
\description{
Parameter Estimates for twostage objects
}
