% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_summary_and_print_methods.R
\name{print.SummaryTwostage}
\alias{print.SummaryTwostage}
\title{Print method for SummaryTwostage objects}
\usage{
\method{print}{SummaryTwostage}(x, ...)
}
\arguments{
\item{x}{An object of class 'SummaryTwostage'}

\item{...}{Additional arguments (currently not used)}
}
\description{
Print method for SummaryTwostage objects
}
